+++ src/vivarium_gates_nutrition_optimization_child/components/__init__.py
@@ -22,6 +22,7 @@ from vivarium_gates_nutrition_optimization_child.components.observers import (
 from vivarium_gates_nutrition_optimization_child.components.population import (
     PopulationLineList,
 )
+from vivarium_gates_nutrition_optimization_child.components.risk import ChildUnderweight
 from vivarium_gates_nutrition_optimization_child.components.wasting import (
     ChildWasting,
     WastingTreatment,