+++ src/vivarium_gates_nutrition_optimization_child/data/loader.py
@@ -94,7 +94,7 @@ def get_data(lookup_key: str, location: str) -> pd.DataFrame:
         data_keys.LRI.CSMR: load_lri_csmr,
         data_keys.LRI.RESTRICTIONS: load_metadata,
         data_keys.MALARIA.DURATION: load_duration,
-        data_keys.MALARIA.PREVALENCE: load_malaria_prevalence,
+        data_keys.MALARIA.PREVALENCE: load_prevalence_from_incidence_and_duration,
         data_keys.MALARIA.INCIDENCE_RATE: load_standard_gbd_2019_data_as_gbd_2021_data,
         data_keys.MALARIA.REMISSION_RATE: load_malaria_remission_rate_from_duration,
         data_keys.MALARIA.DISABILITY_WEIGHT: load_standard_gbd_2019_data_as_gbd_2021_data,
@@ -113,6 +113,8 @@ def get_data(lookup_key: str, location: str) -> pd.DataFrame:
         data_keys.STUNTING.EXPOSURE: load_gbd_2021_exposure,
         data_keys.STUNTING.RELATIVE_RISK: load_gbd_2021_rr,
         data_keys.STUNTING.PAF: load_categorical_paf,
+        data_keys.UNDERWEIGHT.EXPOSURE: load_underweight_exposure,
+        data_keys.UNDERWEIGHT.CATEGORIES: load_metadata,
         data_keys.PEM.EMR: load_pem_emr,
         data_keys.PEM.CSMR: load_pem_csmr,
         data_keys.PEM.RESTRICTIONS: load_pem_restrictions,
@@ -329,6 +331,7 @@ def load_prevalence_from_incidence_and_duration(key: str, location: str) -> pd.D
         cause = {
             data_keys.DIARRHEA.PREVALENCE: data_keys.DIARRHEA,
             data_keys.LRI.PREVALENCE: data_keys.LRI,
+            data_keys.MALARIA.PREVALENCE: data_keys.MALARIA,
         }[key]
     except KeyError:
         raise ValueError(f"Unrecognized key {key}")
@@ -348,22 +351,6 @@ def load_prevalence_from_incidence_and_duration(key: str, location: str) -> pd.D
     # If cause is diarrhea, set early and late neonatal groups prevalence to that of post-neonatal age group
     if key == data_keys.DIARRHEA.PREVALENCE:
         prevalence = utilities.scrub_neonatal_age_groups(prevalence)
-    return prevalence
-
-
-def load_malaria_prevalence(key: str, location: str) -> pd.DataFrame:
-    """Get standard prevalence but update early neonatal data to be
-    (birth_prevalence + prevalence_cause) / 2
-    where birth_prevalence is 0 and prevalence_cause is cause prevalence from GBD.
-    """
-    prevalence = load_standard_gbd_2019_data_as_gbd_2021_data(key, location)
-    birth_prevalence = data_values.BIRTH_PREVALENCE_OF_ZERO
-    enn_prevalence = prevalence.query("age_start == 0")
-    enn_prevalence = (birth_prevalence + enn_prevalence) / 2
-    all_other_prevalence = prevalence.query("age_start != 0.0")
-
-    prevalence = pd.concat([enn_prevalence, all_other_prevalence]).sort_index()
-
     return prevalence
 
 
@@ -415,6 +402,42 @@ def load_emr_from_csmr_and_prevalence(key: str, location: str) -> pd.DataFrame:
     if key == data_keys.DIARRHEA.EMR:
         data.loc[data.index.get_level_values("age_start") < metadata.NEONATAL_END_AGE, :] = 0
     return data
+
+
+def load_underweight_exposure(key: str, location: str) -> pd.DataFrame:
+    """Read in exposure distribution data (conditional on stunting
+    and wasting) from file and transform. This data looks like standard
+    categorical exposure distribution data but with stunting and wasting
+    parameter values in the index."""
+    df = pd.read_csv(paths.UNDERWEIGHT_CONDITIONAL_DISTRIBUTIONS)
+    # add early neonatal data by copying late neonatal
+    early_neonatal = df[df["age_group_name"] == "late_neonatal"].copy()
+    early_neonatal["age_group_name"] = "early_neonatal"
+    df = pd.concat([early_neonatal, df])
+
+    # add age start and age end data instead of age group name
+    age_bins = get_data(data_keys.POPULATION.AGE_BINS, location).reset_index()
+    age_bins["age_group_name"] = age_bins["age_group_name"].str.lower().str.replace(" ", "_")
+    age_start_map = dict(zip(age_bins["age_group_name"], age_bins["age_start"]))
+    age_end_map = dict(zip(age_bins["age_group_name"], age_bins["age_end"]))
+    df["age_start"] = df["age_group_name"].map(age_start_map)
+    df["age_end"] = df["age_group_name"].map(age_end_map)
+    df = df.drop("age_group_name", axis=1)
+
+    # duplicate data for 1990 to 2019
+    year_list = list(range(1990, 2020))
+    years = pd.DataFrame({"year_start": year_list}).set_index(pd.Index([1] * len(year_list)))
+    df = df.set_index(pd.Index([1] * len(df))).join(years)
+    df["year_end"] = df["year_start"] + 1
+
+    # define index
+    df = df.rename({"underweight_parameter": "parameter"}, axis=1)
+    df = df.set_index(
+        metadata.ARTIFACT_INDEX_COLUMNS
+        + ["stunting_parameter", "wasting_parameter", "parameter"]
+    )
+
+    return df.sort_index()
 
 
 def load_gbd_2021_exposure(key: str, location: str) -> pd.DataFrame: