+++ CHANGELOG.rst
@@ -1,3 +1,20 @@
+**v0.3.1 - 09/01/23**
+
+ - Update malaria prevalence to be calculated instead of taken from GBD
+
+**v0.4.0 - 09/01/23**
+
+ - Add underweight exposure
+
+**v0.3.0 - 08/30/23**
+
+ - Include malaria
+
+**v0.2.0 - 08/29/23**
+
+ - Add Dynamic Child Wasting Model with GBD 2021 data
+ - Re-bin to 2021 age groups 
+
 **v0.1.1 - 08/24/23**
 
  - Include effects of antenatal supplementation on gestational age